new font, slight ui improvements/consistency updates
