+++ ui/page/event-list.page.ts
@@ -2,6 +2,7 @@ import { EventModel } from "../../server/events/event.model";
 import { EventList } from "../component/event-list.component";
 import { Column, Heading2 } from "../core/html.elements";
 import { Page, PageModel } from "../core/html.interfaces";
+import { Inline } from "../core/html.operator";
 import { Layout } from "../_Layout";
 
 interface EventListModel extends PageModel {
@@ -15,8 +16,11 @@ export const EventListPage: Page<EventListModel> = (model: EventListModel) => {
         Column(
             Heading2("Check best rated events:"),
             EventList(model.bestRatedEvents),
-            Heading2("Or get hyped for upcoming ones:"),
-            EventList(model.upcomingEvents),
+            model.upcomingEvents.length > 0
+                ? Inline(
+                    Heading2("Or get hyped for upcoming ones:"),
+                    EventList(model.upcomingEvents),
+                ) : null,
         )
     )
