+++ ui/page/sport-details.page.ts
@@ -3,7 +3,7 @@ import { UserModel } from "../../server/users/users.model";
 import { EventForm } from "../component/event-form.component";
 import { EventList } from "../component/event-list.component";
 import { SportComponent } from "../component/sport.component";
-import { Column } from "../core/html.elements";
+import { Column, Heading3 } from "../core/html.elements";
 import { Page, PageModel } from "../core/html.interfaces";
 import { Layout } from "../_Layout";
 
@@ -20,6 +20,7 @@ export const SportDetailsPage: Page<SportDetailsModel> = (model: SportDetailsMod
                 sport: model.sport, 
                 userData: model.userData
             }),
+            Heading3("Events:"),
             EventList(model.sport.events),
             EventForm({
                 sportId: model.sport.id,
